feat : #46 AI리뷰 답글 생성 기능 완료
AI리뷰 답글 생성기능 구현 완료 하였으나 프롬프팅을 더 진행할 예정

+++ src/main/java/com/example/backend/entity/comment/Reply.java
@@ -23,7 +23,7 @@ import lombok.Data;
 public class Reply {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;  // 리뷰의 고유 ID
+    private Long id;  // 리뷰의 고유 ID
 
     @Column(nullable = false)
     private String comment;  // 리뷰 내용
